Move protobuf out of encoding module
