+++ examples/timestamping/backend/src/proto/mod.rs
@@ -21,4 +21,4 @@ pub use self::timestamping::{Timestamp, TimestampEntry, TxTimestamp};
 
 include!(concat!(env!("OUT_DIR"), "/protobuf_mod.rs"));
 
-use exonum::encoding::protobuf::*;
+use exonum::proto::*;
