+++ exonum/src/proto/mod.rs
