+++ exonum/src/proto/schema/exonum/storage.proto
