+++ exonum/src/sandbox/timestamping.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub use encoding::protobuf::tests::TimestampTx;
+pub use proto::tests::TimestampTx;
 
 use rand::{RngCore, SeedableRng};
 use rand_xorshift::XorShiftRng;
