+++ exonum/tests/explorer/blockchain/proto/tests_transactions.proto
@@ -14,7 +14,7 @@
 
 syntax = "proto3";
 
-import "helpers.proto";
+import "exonum/helpers.proto";
 
 package exonum.tests.explorer;
