+++ exonum_build/src/lib.rs
@@ -103,7 +103,7 @@ fn generate_mod_rs<P: AsRef<Path>, Q: AsRef<Path>>(
 /// include!(concat!(env!("OUT_DIR"), "/example_mod.rs"));
 ///
 /// // If you use types from `exonum` .proto files.
-/// use exonum::encoding::protobuf::*;
+/// use exonum::proto::*;
 /// ```
 pub fn protobuf_generate<P, R, I, T>(input_dir: P, includes: I, mod_file_name: T)
 where
